+++ packages/components/pages/f-loyalty/src/components/HowItWorks.vue
@@ -113,8 +113,9 @@
 <script>
 import FCard from '@justeat/f-card';
 import MediaElement from '@justeat/f-media-element';
-
-const TEMP_IMG_PATH = 'https://d30v2pzvrfyzpo.cloudfront.net/b/hw/img/decoration/';
+import { globalisationServices } from '@justeat/f-services';
+import tenantConfigs from '../tenants';
+import tenantBagFees from '../tenantBagFees';
 
 const formatAsCurrency = (amount, currencySymbol) => `${currencySymbol}${amount
     .toFixed(2)}`;
@@ -127,21 +128,25 @@ export default {
         MediaElement
     },
 
-    data: () => ({
-        bagsData: [
-            40,
-            50,
-            45,
-            40,
-            55
-        ],
-        flexLayout: {
-            default: {
-                column: true,
-                reverse: true
-            }
+    props: {
+        locale: {
+            type: String,
+            default: ''
         }
-    }),
+    },
+
+    data () {
+        const locale = globalisationServices.getLocale(tenantConfigs, this.locale, this.$i18n);
+        return {
+            tenant: locale,
+            flexLayout: {
+                default: {
+                    column: true,
+                    reverse: true
+                }
+            }
+        };
+    },
 
     computed: {
         bags () {
@@ -152,48 +157,47 @@ export default {
             // _Very_ naive way of implementing ordinals but will suit our purposes here
             const ordinal = { 0: 'st', 1: 'nd', 2: 'rd' };
 
-            return this.bagsData.map((bagValue, index) => ({
+            return [...tenantBagFees(this.tenant)].map((bagValue, index) => ({
                 number: `${index + 1}${ordinal[index] || 'th'}`,
-                image: `${TEMP_IMG_PATH}${bagValue}-small-object-bag-${currency}.svg`,
+                image: `${this.$t('howItWorks.imagePath')}bag-${currency}-${bagValue}.svg`,
                 value: formatAsCurrency(bagValue * percentage * 0.01, currencySymbol)
                     .replace('.00', '') // remove zero pence/cents amounts for bag values
             }));
         },
 
         mediaCards () {
-            const percentage = this.$t('percentage');
-
             return [
                 {
                     title: this.$t('howItWorks.media.cards.order.title'),
                     text: this.$t('howItWorks.media.cards.order.text'),
-                    image: `${TEMP_IMG_PATH}stampcards-complex-object-restaurant-promo.svg`
+                    image: this.$t('howItWorks.media.cards.order.image')
                 },
                 {
                     title: this.$t('howItWorks.media.cards.earn.title'),
                     text: this.$t('howItWorks.media.cards.earn.text'),
-                    image: `${TEMP_IMG_PATH}stampcards-complex-object-login-promo-full-colour-${percentage}.svg`
+                    image: this.$t('howItWorks.media.cards.earn.image')
                 },
                 {
                     title: this.$t('howItWorks.media.cards.collect.title'),
                     text: this.$t('howItWorks.media.cards.collect.text'),
-                    image: `${TEMP_IMG_PATH}stampcards-complex-object-stampcard-colour-03-${percentage}-full.svg`
+                    image: this.$t('howItWorks.media.cards.collect.image')
                 },
                 {
                     title: this.$t('howItWorks.media.cards.reward.title'),
                     text: this.$t('howItWorks.media.cards.reward.text'),
-                    image: `${TEMP_IMG_PATH}stampcards-bag-order-refund.svg`
+                    image: this.$t('howItWorks.media.cards.reward.image')
                 }
             ];
         },
 
         total () {
             return formatAsCurrency(
-                this.bagsData.reduce((sum, current) => sum + current, 0) * this.$t('percentage') * 0.01,
+                [...tenantBagFees(this.tenant)].reduce((sum, current) => sum + current, 0) * this.$t('percentage') * 0.01,
                 this.$t('currencySymbol')
             );
         }
     }
+
 };
 </script>
 
@@ -352,12 +356,13 @@ HOW IT WORKS SECTION: EXAMPLE SECTION
 
 }
 
-@include f.media('<=narrowMid') {
-    .c-howItWorks-example-order-image {
+.c-howItWorks-example-order-image {
+    width: 88px;
+
+    @include f.media('<=narrowMid') {
         width: 45px;
     }
 }
-
 
 .c-howItWorks-example-percentages {
     display: flex;