+++ packages/components/pages/f-loyalty/src/components/_tests/HowItWorks.test.js
@@ -24,8 +24,9 @@ beforeEach(() => {
 describe('HowItWorksLayout', () => {
     it('should be defined', () => {
         // Arrange
+        const locale = 'en-AU';
         const i18n = {
-            locale: 'en-AU',
+            locale,
             fallbackLocale: 'en-AU',
             messages
         };
@@ -34,6 +35,9 @@ describe('HowItWorksLayout', () => {
         wrapper = shallowMount(HowItWorksLayout, {
             localVue,
             i18n,
+            propsData: {
+                locale
+            },
             mocks: {
                 $t: key => i18nMocker(key, i18n.locale)
             }
@@ -46,8 +50,9 @@ describe('HowItWorksLayout', () => {
     describe('when in AU', () => {
         beforeEach(() => {
             // Arrange
+            const locale = 'en-AU';
             const i18n = {
-                locale: 'en-AU',
+                locale,
                 fallbackLocale: 'en-AU',
                 messages
             };
@@ -56,6 +61,9 @@ describe('HowItWorksLayout', () => {
             wrapper = shallowMount(HowItWorksLayout, {
                 localVue,
                 i18n,
+                propsData: {
+                    locale
+                },
                 mocks: {
                     $t: key => i18nMocker(key, i18n.locale)
                 }
@@ -67,27 +75,27 @@ describe('HowItWorksLayout', () => {
             expect(wrapper.vm.bags).toEqual([
                 {
                     number: '1st',
-                    image: 'https://d30v2pzvrfyzpo.cloudfront.net/b/hw/img/decoration/40-small-object-bag-dollar.svg',
+                    image: 'https://just-eat-prod-sg-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-dollar-40.svg',
                     value: '$6'
                 },
                 {
                     number: '2nd',
-                    image: 'https://d30v2pzvrfyzpo.cloudfront.net/b/hw/img/decoration/50-small-object-bag-dollar.svg',
+                    image: 'https://just-eat-prod-sg-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-dollar-50.svg',
                     value: '$7.50'
                 },
                 {
                     number: '3rd',
-                    image: 'https://d30v2pzvrfyzpo.cloudfront.net/b/hw/img/decoration/45-small-object-bag-dollar.svg',
+                    image: 'https://just-eat-prod-sg-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-dollar-45.svg',
                     value: '$6.75'
                 },
                 {
                     number: '4th',
-                    image: 'https://d30v2pzvrfyzpo.cloudfront.net/b/hw/img/decoration/40-small-object-bag-dollar.svg',
+                    image: 'https://just-eat-prod-sg-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-dollar-40.svg',
                     value: '$6'
                 },
                 {
                     number: '5th',
-                    image: 'https://d30v2pzvrfyzpo.cloudfront.net/b/hw/img/decoration/55-small-object-bag-dollar.svg',
+                    image: 'https://just-eat-prod-sg-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-dollar-55.svg',
                     value: '$8.25'
                 }
             ]);
@@ -102,16 +110,21 @@ describe('HowItWorksLayout', () => {
     describe('when in UK', () => {
         beforeEach(() => {
             // Arrange
+            const locale = 'en-GB';
             const i18n = {
-                locale: 'en-GB',
+                locale,
                 fallbackLocale: 'en-GB',
                 messages
             };
+
 
             // Act
             wrapper = shallowMount(HowItWorksLayout, {
                 localVue,
                 i18n,
+                propsData: {
+                    locale
+                },
                 mocks: {
                     $t: key => i18nMocker(key, i18n.locale)
                 }
@@ -123,27 +136,27 @@ describe('HowItWorksLayout', () => {
             expect(wrapper.vm.bags).toEqual([
                 {
                     number: '1st',
-                    image: 'https://d30v2pzvrfyzpo.cloudfront.net/b/hw/img/decoration/40-small-object-bag-pound.svg',
+                    image: 'https://just-eat-prod-eu-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-pound-40.svg',
                     value: '£4'
                 },
                 {
                     number: '2nd',
-                    image: 'https://d30v2pzvrfyzpo.cloudfront.net/b/hw/img/decoration/50-small-object-bag-pound.svg',
+                    image: 'https://just-eat-prod-eu-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-pound-50.svg',
                     value: '£5'
                 },
                 {
                     number: '3rd',
-                    image: 'https://d30v2pzvrfyzpo.cloudfront.net/b/hw/img/decoration/45-small-object-bag-pound.svg',
+                    image: 'https://just-eat-prod-eu-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-pound-45.svg',
                     value: '£4.50'
                 },
                 {
                     number: '4th',
-                    image: 'https://d30v2pzvrfyzpo.cloudfront.net/b/hw/img/decoration/40-small-object-bag-pound.svg',
+                    image: 'https://just-eat-prod-eu-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-pound-40.svg',
                     value: '£4'
                 },
                 {
                     number: '5th',
-                    image: 'https://d30v2pzvrfyzpo.cloudfront.net/b/hw/img/decoration/55-small-object-bag-pound.svg',
+                    image: 'https://just-eat-prod-eu-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-pound-55.svg',
                     value: '£5.50'
                 }
             ]);
@@ -152,6 +165,126 @@ describe('HowItWorksLayout', () => {
         it('total should be calculated as expected', () => {
             // Assert
             expect(wrapper.vm.total).toEqual('£23.00');
+        });
+    });
+
+    describe('when in NZ', () => {
+        beforeEach(() => {
+            // Arrange
+            const locale = 'en-NZ';
+            const i18n = {
+                locale,
+                fallbackLocale: 'en-NZ',
+                messages
+            };
+
+            // Act
+            wrapper = shallowMount(HowItWorksLayout, {
+                localVue,
+                i18n,
+                propsData: {
+                    locale
+                },
+                mocks: {
+                    $t: key => i18nMocker(key, i18n.locale)
+                }
+            });
+        });
+
+        it('bags should be calculated as expected', () => {
+            // Assert
+            expect(wrapper.vm.bags).toEqual([
+                {
+                    number: '1st',
+                    image: 'https://just-eat-prod-sg-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-dollar-40.svg',
+                    value: '$4'
+                },
+                {
+                    number: '2nd',
+                    image: 'https://just-eat-prod-sg-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-dollar-50.svg',
+                    value: '$5'
+                },
+                {
+                    number: '3rd',
+                    image: 'https://just-eat-prod-sg-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-dollar-45.svg',
+                    value: '$4.50'
+                },
+                {
+                    number: '4th',
+                    image: 'https://just-eat-prod-sg-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-dollar-40.svg',
+                    value: '$4'
+                },
+                {
+                    number: '5th',
+                    image: 'https://just-eat-prod-sg-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-dollar-55.svg',
+                    value: '$5.50'
+                }
+            ]);
+        });
+
+        it('total should be calculated as expected', () => {
+            // Assert
+            expect(wrapper.vm.total).toEqual('$23.00');
+        });
+    });
+
+    describe('when in IE', () => {
+        beforeEach(() => {
+            // Arrange
+            const locale = 'en-IE';
+            const i18n = {
+                locale,
+                fallbackLocale: 'en-IE',
+                messages
+            };
+
+            // Act
+            wrapper = shallowMount(HowItWorksLayout, {
+                localVue,
+                i18n,
+                propsData: {
+                    locale
+                },
+                mocks: {
+                    $t: key => i18nMocker(key, i18n.locale)
+                }
+            });
+        });
+
+        it('bags should be calculated as expected', () => {
+            // Assert
+            expect(wrapper.vm.bags).toEqual([
+                {
+                    number: '1st',
+                    image: 'https://just-eat-prod-eu-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-euro-25.svg',
+                    value: '€2.50'
+                },
+                {
+                    number: '2nd',
+                    image: 'https://just-eat-prod-eu-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-euro-30.svg',
+                    value: '€3'
+                },
+                {
+                    number: '3rd',
+                    image: 'https://just-eat-prod-eu-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-euro-35.svg',
+                    value: '€3.50'
+                },
+                {
+                    number: '4th',
+                    image: 'https://just-eat-prod-eu-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-euro-25.svg',
+                    value: '€2.50'
+                },
+                {
+                    number: '5th',
+                    image: 'https://just-eat-prod-eu-res.cloudinary.com/image/upload/v1630068495/Experiments/Homeweb-Coreweb/bag-euro-30.svg',
+                    value: '€3'
+                }
+            ]);
+        });
+
+        it('total should be calculated as expected', () => {
+            // Assert
+            expect(wrapper.vm.total).toEqual('€14.50');
         });
     });
 });