+++ packages/components/pages/f-loyalty/CHANGELOG.md
@@ -4,6 +4,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v2.6.0
+------------------------------
+*November 7, 2022*
+
+### Changed
+- Updated tenant files to include new image paths
+- Updated tenant files to correct percentages
+- Updated test file to have new tenants tested
+
 v2.5.0
 ------------------------------
 *August 3, 2022*